+++ NEWS.md
@@ -5,6 +5,8 @@
 * Keep signal to noise data when importing/exporting XCMS data (`sn` column) (suggested by Ricardo Cunha)
 * Reversed argument order of `exportedData`/`verbose` to `getXCMSSet()` functions to avoid ambiguities
 * Automated tests for importing/exporting XCMS(3) data + small fixes for surfaced bugs
+* `generateComponentsNontarget()`: allow wider _m/z_ deviation for proper linkage of different series (controlled by `absMzDevLink` argument).
+
 
 # patRoon 1.0.1